Drop all but first line during location parsing
This is a fix but not really a good solution for #63. The problem arises
because the raw format (at least in go1.9) can contain lines of the form

```
[...]
   729: 0x4d6a395 github.com/cockroachdb/cockroach/pkg/sql.(*tableState).release /Users/tschottdorf/go/src/github.com/cockroachdb/cockroach/pkg/sql/lease.go:791 s=0
   730: 0x4021625 runtime.heapBits.next /usr/local/Cellar/go/1.9beta2/libexec/src/runtime/mbitmap.go:464 s=0
             runtime.scanobject /usr/local/Cellar/go/1.9beta2/libexec/src/runtime/mgcmark.go:1162 s=0
[...]
```

and the last line is not parsed correctly. The way parsing is set up is
line-by-line, so it's not straightforward to recover. Instead of a bigger
refactor to try to address this, this change just drops these extra lines,
and we should really implement #64.

Note also that I changed some lines in the parsing test to match what is
actually produced by go1.9 -- I'm not sure if the previous snippet was
simply not true-to-original or whether perhaps the raw format has changed.
Either way, using the raw format seems to invite regular breakage.

Closes #63.